Improved reading unknown escape sequences
* Ending in modes other than "outside" left the escape seqences that you
are in the middle of reading unprinted (though the "count" chars had
accounted for them)
* Tiny typing issue when you are writing the
firstCsiChar/secondCsiChar/ch; it needed to be []byte.
* The counting just seems to be off for these new cases, one case where
this happens is the escape seqence \x1b= which is used by xterm to use
DECPATH mode.   The idea then would be that when you start reading a
sequence, but then later leave it (firstCsiCode -> outsideCsiCode, maybe
another transition too) you need to adjust the count so that it includes
them only as it actually writes the characters and not in the nc
variable.